requirements.txt
pillow
zoneinfo
